pyproject.toml
[project]
name = "ai-driven-plan-engine"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "cairosvg>=2.8.2",
    "ortools>=9.10.4067",
    "pyyaml>=6.0.1",
    "svgwrite>=1.4.3",
]

[dependency-groups]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.3.4",
]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
extend-select = [
    "E",
    "W",    # pycodestyle
    "F",    # Pyflakes
    "UP",   # pyupgrade
    "N",    # pep8-naming
    "C90",  # complexity
    "C4",
    "PIE",
    "T20",
    "SIM",
    "ERA",
    "FIX",
    "INT",
    "TCH",
    "TID",
    "SLOT",
    "RSE",
    "Q",
    "PYI",
    "INP",
    "B",
    "PTH",
    "SLF",
    "PT",
    "RET",
    "I",    # isort
    "PL",   # pylint
    "RUF",  # ruff
]
ignore = [
    "E501",     # line too long
    "PLR2004",  # magic value comparison
    "PLR0914",  # too many local variables
    "PLR0913",  # too many arguments
    "PLR0912",  # too many branches
    "PLR0911",  # too many return statements
    "RUF001",   # ambiguous characters
    "FIX002",   # found TODO
    "T201",     # print found
    "SLF001",   # private member accessed (renderer sub-modules share _x/_y)
    "PLR0915",  # too many statements
]

[tool.ruff.lint.isort]
known-first-party = ["plan_engine"]

[tool.ruff.lint.mccabe]
max-complexity = 30

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["plan_engine"]

[tool.coverage.report]
fail_under = 80
